requirements.txt
asgiref==3.4.1
Django==3.2.8
django-environ==0.7.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
environ==1.0
psycopg2==2.9.1
PyJWT==2.3.0
python-json-logger==2.0.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
